Solution to Programming Assignment 2
Created two functions as required. Function 1 - Creates a list
consisting of functions. Function 2 - Checks if the inverse of matrix
already exists and value of matrix itself has not change. If both are
true, then returns the value of inverse from cache. If not, then
calculates the value of inverse for given matrix.